requirements.txt
streamlit==1.32.0
langchain==0.0.285
chromadb==0.4.15
openai==0.28.1
python-dotenv==1.0.0
tiktoken==0.5.2
unstructured==0.10.30
pdf2image==1.16.3
python-docx==0.8.11
openpyxl==3.1.2
pandas==2.0.3
PyPDF2==3.0.1